rename 'server_api' to '_api'
